Change header nunjucks template
Update nunjucks template to make it easier to integrate the header into
prototypes using the Prototype Kit. Now the entire header NJK file can
be copied into prototypes layout template. Navigation links and the name
of the service can be set in the view, using the set syntax in Nunjucks,
e.g. {% set serviceName = 'Apply for a juggling licence' %}