GUA-854: Enlarge target area for service nav links
There was some feedback suggesting improvements for the focus state on
the service links on the One Login header.

The feedback was reviewed by UCD and there was agreement that the target
area for the links should be increased to take up the entire height of
the nav bar in the desktop variation of the header.

This eliminates the unsigtly double border when an active link is
focused.