GUA-899: Optimisations for prototype kit/npm
Implement configuration to make it easier for users to install the
header in their Prototype Kit projects. This configuration specifies
load paths for the SCSS and JS as well as Nunjucks paths making it so
that the Nunjucks template for the header can be included easily into
any Prototype Kit project.

There is an additional step to have the header added as a plugin to the
Prototype Kit UI which would allow users to install it by simply
clicking a button. For this to happen the header must be manually
added to an allow list of plugins by someone from the Prototype Kit
team.

Organise the `dist` folder based on file type