BAU: Fix deploy to GH Pages action
The deploy to Github pages action is failing after adding the
documentation on installing the header in the Prototype Kit.

This is because GH pages uses Jekyll to convert MD into HTML, which uses
Liquid as a templating language, which has a very similar syntax to
Nunjucks.
The {% %} tags used to exemplify usage of the header in
prototype-kit-installation.md are picked up as Liquid code, causing
errors in the deploy to pages action.

This is a workaround which uses Liquid's {% raw %}...{% endraw %} tags
to escape the Nunjucks/Liquid syntax in the documentation file which
should eliminate the build errors.