OLH-1031: Fix deploy to Github Pages (hopefully)
The deploy to GH Pages action is failing since adding the documentation
on how to install the header with the Prototype Kit. This is because GH
Pages uses Jekyll to convert .md into .html, which uses Liquid as a
templating language, which has a very similar syntax to Nunjucks.

The snippets of Nunjucks language in our documentation are picked up as
Liquid code, causing errors in the deploy to Pages, as Liquid does not
have the same syntax as Nunjucks so certain keywords such as `set` are
not recognised.

This replaces some of the Nunjucks syntax in our documentation (the {}
and % characters) with HTML entities which should not be recognised as
Liquid code. It also swaps the Markdown syntax for code snippets with
actual HTML tags, as it appears HTML entities don't render correctly
inside `...` and ```...``` blocks.