OLH-1159: Prevent uncaught type error
Currently if someone copies both service-header.js and
init-service-header.js from us instead of writing their own conditional
initialisation JS, and includes both on every page of their site
including pages which do not contain the OL header, there is a JS error
on pages where the header is not present.
This fixes that issue.