OLH-1159: slight CSS tweak for service name
Tweak the CSS so that the service name is better spaced in the OL header
in the mobile variation in the absence of a service navigational menu.
When there is a service nav menu, it should look exactly the same as
before.
When the nav displays with service name only, the visual spacing should
now be more balanced.