OLH-1043: Tweak service nav for 1 link scenario
Some services might have only one service link in the service
navigation. In this scenario it does not make sense to collapse
navigation into a dropdown in the mobile variant.

This tweaks the way the nav script works so that dropdown behaviour does
not initialise unless at least 2 navigational links are present.