Support overridding sign out link on PK layout
Support over-writing the default Sign Out link when using the "Page
with GOVUK One Login header" template for the prototype kit.

The value of this link can be over ruled by passing a custom
signOutLink value into the govukOneLoginServiceHeader macro so users
importing the macro directly would be able to set a custom value this
way.

Users using the pre-built template would be unable to set a custom
signOutLink. This update should fix that.