OLH-2363: Add doc for devs working on header
Add documentation aimed at internal developers working on the header
component. The documentation outlines noteworthy information about the
setup on the repo as well as guidance for versioning and changelog
maintenance.

Add github PR template with checklist of reminders – ideally some of the
manual pre-commit tasks should be automated but this should do for now