BAU: Refactor + a11y optimisations (CSS)
Remove unneeded mixins (previously needed because some of the styles for
the OL menu and the service menu were the same. The service menu styles
come from the DS now.

Toggle visibility on `<ul>` container instead of the `<nav>` when the
OL menu toggle button is interacted with.
It was raised that using `display: none` on the nav removes it from the
a11y tree which is not accessible for AT users.