BAU: Update main parameter in package.json
The value of the main parameter in package.json was set to the default
index.js which does not exists. This update makes it point to the
service header JS file which should make it easier to
import/initialise the script using ES6 import syntax by consumers who
choose this approach.