BAU: forced colours border tweak
A transparent border is applied to the header for readability in forced
colours mode. This border is, however, causing a minor layout issue when
the dropdown toggle is activated on mobile, where the button looks like
it's shifting by 1 px. This ensures the border is only applied when the
forced colours media query is active.

It doesn't seem to have a detrimental effect otherwise – it's actually
bringing it closer to the original intended design as the border
dividing the OL nav and secondary navs is now the intended 1px vs 2px.