BAU: bump govuk-frontend to 5.11.1
Update logo SVG to latest one provided in the release notes for 5.11.0.
There is no visual difference unless very zoomed in (my understanding
is the logo is supposed to look slightly sharper as a result of this
change!).

The service header component was extended with a
collapseNavigationOnMobile nunjucks option so the option has been added
to the list of parameters accepted in the prototype kit layout.