OLH-2940: Minor spacing tweaks
Implement tweaks to bring the service header more in line with the
GOV.UK/Design System header. The spacing on the homepage link has been
adjusted so that the height and alignment of the logo aligns more
closely with the rebranded DS version.
Additionally, the height of the nav has been tweaked as it was off by
1px. I have done this by adding a border which brings the added bonus of
the header displaying better in forced colours mode.
Another tweak has been the toggle button which has taller and had
slightly different spacing to the GOVUK example.