BAU: Hide GOVUK OL icon SVG from screen readers
The "OL sideways" icon is purely decorative in the context of the
cross-service header.
Labelling it with alt text or with a title would add noise for AT users,
as the visual element is already accompanied by descriptive text
("GOV.UK One Login").
Adding `aria-hidden` ensures the element is hidden for AT. The `role`
attribute became redundant in this context so it's been removed.