OLH-3183: Remove support for old (black bar) header
When the updates supporting the new rebranded version of the GOVUK
header were introduced, the code for the old header was not removed
immediately. We did this to facilitate testing and delivery for RPs,
as well as rollback in case of launch not going as planned.
The new header design has been in use for a while, so support for the
 old header can be removed as there'd be no rolling back to the old
design at this stage.