OLH-3349: Support Sign out as POST request and minor a11y tweak
1. Extend service header with options to specify a form action and form
inputs for services opting to implement their logout functionality as
a POST. This was implemented in response to a service raising an issue/
requesting support with this matter.
2. Implement accessibility optimisation for high contrast mode. The
Sign out and GOVUK One Login links in the header did not display with
an underline on hover when tested with Chrome's forced-colors emulator.
This should be fixed now.